chore: remove sort option
